remove PluginBrige and all code which becomes unreachable by doing so
