+++ common/src/main/java/net/adoptopenjdk/icedteaweb/commandline/CommandLineOptionsDefinition.java
@@ -93,7 +93,6 @@ public class CommandLineOptionsDefinition {
                 CommandLineOptions.OFFLINE,
                 CommandLineOptions.TRUSTNONE,
                 CommandLineOptions.JNLP,
-                CommandLineOptions.HTML,
                 CommandLineOptions.BROWSER);
     }
