+++ core/src/main/java/net/sourceforge/jnlp/cache/ResourceDownloader.java
@@ -201,12 +201,7 @@ public class ResourceDownloader implements Runnable {
                 if (jnlpPath == null || jnlpPath.equals("")) {
                     jnlpPath = Boot.getOptionParser().getParam(CommandLineOptions.JNLP);
                     if (jnlpPath == null || jnlpPath.equals("")) {
-                        jnlpPath = Boot.getOptionParser().getParam(CommandLineOptions.HTML);
-                        if (jnlpPath == null || jnlpPath.equals("")) {
-                            LOG.info("Not-setting jnlp-path for missing main/jnlp/html argument");
-                        } else {
-                            entry.setJnlpPath(jnlpPath);
-                        }
+                        LOG.info("Not-setting jnlp-path for missing main/jnlp argument");
                     } else {
                         entry.setJnlpPath(jnlpPath);
                     }
